MAJ des commentaires  - MAJ de la route createSauce

+++ backend/controllers/sauce.js
@@ -1,31 +1,46 @@
+// Importation du Model sauce
 const Sauce = require('../models/sauce')
 const fs = require("fs")
 
+// Exportation de la fonction pour la création d'une sauce
 exports.createSauce = (req, res, next) => {
     const sauceObject = JSON.parse(req.body.sauce)
    // Supression des données non souhaitées
     delete sauceObject._id
-    delete sauceObject._userId
     delete sauceObject.likes
     delete sauceObject.dislikes
-
-    const sauce = new Sauce({
-        ...sauceObject,
-        userId: req.auth.userId,
-        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
-    })
-    sauce.save()
-        .then(() => res.status(201).json({ message: "Objet enregistré !" }))
-        .catch(error => {res.status(400).json({ error });
+    // Vérification de la valeur de heat
+    const heat = sauceObject.heat
+    if (heat <= 0 || heat > 10){
+        fs.unlink(`images/${req.file.filename}`,()=>{
+            res.status(400).json({ message: "Mauvaise requete"})
         })
+    }else{
+      // Création d'une instance du model Sauce
+      const sauce = new Sauce({
+          ...sauceObject,
+          userId: req.auth.userId,
+          // Reconstrution de l'URL complète du fichier enregistré
+          imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`,
+          usersLiked : [],
+          usersDisliked: []
+      })
+        // Sauvegarde de la sauce crée dans la DB
+        sauce.save()
+            .then(() => res.status(201).json({ message: "Objet enregistré !" }))
+            .catch(error => {res.status(400).json({ error });
+            })
+    }
 }
 
+// Exportation de la fonction pour récupérer une sauce
 exports.getOneSauce = (req, res, next) => {
     Sauce.findOne({ _id: req.params.id })
         .then(sauce => res.status(200).json(sauce))
         .catch(error => res.status(404).json({ error }))
 }
 
+// Exportation de la fonction pour modifier une sauce
 exports.modifySauce = (req, res, next) => {
     const sauceObject = req.file ? {
         ...JSON.parse(req.body.sauce),
@@ -37,16 +52,15 @@ exports.modifySauce = (req, res, next) => {
         .then(sauce => {
             if (sauce.userId != req.auth.userId) {
                 res.status(401).json({ message: "Non AUTORISER" })
-
             } 
-            if(req.file){
+            if(req.file && sauce.userId === req.auth.userId ){
                 const filename = sauce.imageUrl.split('/images/')[1]
                 fs.unlink(`images/${filename}`, () => {
                     Sauce.updateOne({ _id: req.params.id}, { ...sauceObject, _id: req.params.id})
-                    .then(() => res.status(200).json({message : 'Objet modifié!. Image précedante effacée de la base'}))
+                    .then(() => res.status(200).json({message : 'Objet modifié!. Image précédente effacée de la base'}))
                     .catch(error => res.status(401).json({ error }))
                 })
-                } else {
+                } else if (sauce.userId === req.auth.userId) {
                 Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
                     .then(() => res.status(201).json({ message: "Objet modifié !" }))
                     .catch(error => { res.status(400).json({ error }) })
@@ -56,6 +70,7 @@ exports.modifySauce = (req, res, next) => {
 
 }
 
+// Exportation de la fonction pour supprimer une sauce
 exports.deleteSauce = (req, res, next) => {
     Sauce.findOne({ _id: req.params.id })
         .then(sauce => {
@@ -73,12 +88,14 @@ exports.deleteSauce = (req, res, next) => {
         .catch(error => res.status(500).js({ error }))
 }
 
+// Exportation de la fonction pour récupérer toutes les sauces
 exports.getAllSauce = (req, res, next) => {
     Sauce.find()
         .then(sauces => res.status(200).json(sauces))
         .catch(error => res.status(400).json({ error }))
 }
 
+// Exportation de la fonction pour liker ou disliker une sauce
 exports.likeSauce = (req, res, next) =>{
     const userId = req.auth.userId
     const like = req.body.like
